# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EcoStore Ecommerce Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update user role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # CATEGORIES
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, newest, bestseller, rating]
        - name: inStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /products/{id}/stock:
    patch:
      operationId: updateProductStock
      tags: [products]
      summary: Update product stock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  # CART
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews (optionally filtered by productId)
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewPage"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a product review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # FORM DESIGNER
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, customer]
        avatarUrl:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        provider:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        avatarUrl:
          type: string
    RoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, manager, customer]

    # CATEGORY
...
    Product:
      type: object
      required: [id, name, slug, price, stock, categoryId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        stock:
          type: integer
        sku:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        isFeatured:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ProductInput:
      type: object
      required: [name, slug, price, stock, categoryId]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
        stock:
          type: integer
        sku:
          type: string
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        isFeatured:
          type: boolean
        tags:
          type: array
          items:
            type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
        stock:
          type: integer
        sku:
          type: string
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        isFeatured:
          type: boolean
        tags:
          type: array
          items:
            type: string
    StockUpdate:
      type: object
      required: [stock]
      properties:
        stock:
          type: integer

    # CART
...
    Review:
      type: object
      required: [id, productId, userId, rating, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        userAvatar:
          type: ["string", "null"]
        rating:
          type: integer
        title:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ReviewPage:
      type: object
      required: [data, total, page, limit, averageRating]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        averageRating:
          type: number
    ReviewInput:
      type: object
      required: [productId, rating]
      properties:
        productId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
        body:
          type: string

    # FORM DESIGNER
...