# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoStore Ecommerce Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  slug: string;
  description?: string;
  price: number;
  compareAtPrice?: number;
  stock: number;
  sku?: string;
  imageUrl?: string;
  images?: string[];
  categoryId: number;
  isFeatured?: boolean;
  tags?: string[];
}